+++ index.js
@@ -1,23 +1,24 @@
 import Books from './modules/Books.js';
 
-import {even} from './modules/addbooks.js';
+import { DateTime } from './modules/date.js';
 
-import { DateTime } from './modules/date.js'
-
-import { showhide } from './modules/showhidesection.js'
+import {
+  even,
+  showhide,
+} from './modules/elements.js';
 
 Books.displayall();
 
 document.querySelector('.add').addEventListener('click', even);
 
 const timeDate = document.querySelector('.date');
- 
+
 const dateAndTime = () => {
-    setInterval(() => {
-      const date = (`${DateTime.now().toLocaleString(DateTime.DATETIME_MED)}`);
-      timeDate.innerHTML = date;
-    }, 0);
-  };
+  setInterval(() => {
+    const date = (`${DateTime.now().toLocaleString(DateTime.DATETIME_MED)}`);
+    timeDate.innerHTML = date;
+  }, 0);
+};
 
 dateAndTime();
